agentic_curriculum/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.0
google-generativeai==0.3.2
aiofiles==23.2.1
